removed validation class and added token decoding method
