+++ src/client/BaseAuthClient.ts
@@ -11,7 +11,6 @@ import {
     CryptoProvider,
 } from "@azure/msal-node";
 
-import { TokenValidator } from "../crypto/TokenValidator";
 import { AppSettings } from "../config/AppSettings";
 
 import {
@@ -24,14 +23,12 @@ export abstract class BaseAuthClient {
     appSettings: AppSettings;
     protected msalConfig: Configuration;
     protected msalClient: ConfidentialClientApplication;
-    protected tokenValidator: TokenValidator;
     protected cryptoProvider: CryptoProvider;
     protected logger: Logger;
 
     protected constructor(appSettings: AppSettings, msalConfig: Configuration) {
         this.appSettings = appSettings;
         this.msalConfig = msalConfig;
-        this.tokenValidator = new TokenValidator(this.appSettings, this.msalConfig, this.logger);
         this.cryptoProvider = new CryptoProvider();
         this.logger = new Logger(this.msalConfig.system.loggerOptions, packageName, packageVersion);
         this.msalClient = new ConfidentialClientApplication(this.msalConfig);
