--- test/crypto/TokenValidator.spec.ts
@@ -1,29 +1,0 @@
-/*
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License.
- */
-
-import { Logger } from "@azure/msal-common";
-import { TokenValidator } from "../../src/crypto/TokenValidator";
-import { TEST_CONSTANTS, TEST_APP_SETTINGS, TEST_MSAL_CONFIG } from "../TestConstants";
-import { DEFAULT_LOGGER_OPTIONS } from "../../src/utils/Constants"
-
-describe('Token validator tests', () => {
-
-    const appSettings = TEST_APP_SETTINGS;
-    const msalConfig = TEST_MSAL_CONFIG;
-    const logger = new Logger(DEFAULT_LOGGER_OPTIONS);
-
-    const tokenValidator = new TokenValidator(appSettings, msalConfig, logger);
-    const testTokenPayload = "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIsImtpZCI6IjFMVE16YWtpaGlSbGFfOHoyQkVKVlhlV01xbyJ9.eyJ2ZXIiOiIyLjAiLCJpc3MiOiJodHRwczovL2xvZ2luLm1pY3Jvc29mdG9ubGluZS5jb20vOTEyMjA0MGQtNmM2Ny00YzViLWIxMTItMzZhMzA0YjY2ZGFkL3YyLjAiLCJzdWIiOiJBQUFBQUFBQUFBQUFBQUFBQUFBQUFJa3pxRlZyU2FTYUZIeTc4MmJidGFRIiwiYXVkIjoiNmNiMDQwMTgtYTNmNS00NmE3LWI5OTUtOTQwYzc4ZjVhZWYzIiwiZXhwIjoxNTM2MzYxNDExLCJpYXQiOjE1MzYyNzQ3MTEsIm5iZiI6MTUzNjI3NDcxMSwibmFtZSI6IkFiZSBMaW5jb2xuIiwicHJlZmVycmVkX3VzZXJuYW1lIjoiQWJlTGlAbWljcm9zb2Z0LmNvbSIsIm9pZCI6IjAwMDAwMDAwLTAwMDAtMDAwMC02NmYzLTMzMzJlY2E3ZWE4MSIsInRpZCI6IjkxMjIwNDBkLTZjNjctNGM1Yi1iMTEyLTM2YTMwNGI2NmRhZCIsIm5vbmNlIjoiMTIzNTIzIiwiYWlvIjoiRGYyVVZYTDFpeCFsTUNXTVNPSkJjRmF0emNHZnZGR2hqS3Y4cTVnMHg3MzJkUjVNQjVCaXN2R1FPN1lXQnlqZDhpUURMcSFlR2JJRGFreXA1bW5PcmNkcUhlWVNubHRlcFFtUnA2QUlaOGpZIn0.1AFWW-Ck5nROwSlltm7GzZvDwUkqvhSQpm55TQsmVo9Y59cLhRXpvB8n-55HCr9Z6G_31_UbeUkoz612I2j_Sm9FFShSDDjoaLQr54CreGIJvjtmS3EkK9a7SJBbcpL1MpUtlfygow39tFjY7EVNW9plWUvRrTgVk7lYLprvfzw-CIqw3gHC-T7IK_m_xkr08INERBtaecwhTeN4chPC4W3jdmw_lIxzC48YoQ0dB1L9-ImX98Egypfrlbm0IBL5spFzL6JDZIRRJOu8vecJvj1mq-IUhGt0MacxX8jdxYLP-KUu2d9MbNKpCKJuZ7p8gwTL5B7NlUdh_dmSviPWrw";
-    
-    it('should generate a valid instance', () => {
-        expect(tokenValidator).toBeInstanceOf(TokenValidator);
-    });
-
-    it('should decode a token', () => {
-        const decodedToken = TokenValidator.decodeAuthToken(testTokenPayload);
-        expect(decodedToken.header).toMatchObject(TEST_CONSTANTS.ID_TOKEN_HEADER);
-        expect(decodedToken.payload).toMatchObject(TEST_CONSTANTS.ID_TOKEN_CLAIMS);
-    });
-})
