+++ test/utils/CryptoUtils.spec.ts
@@ -4,6 +4,8 @@
  */
 
 import { CryptoUtils } from '../../src/utils/CryptoUtils';
+import { TEST_CONSTANTS } from "../TestConstants";
+
 
 describe('Crypto utilities tests', () => {
     const cryptoUtils = new CryptoUtils();
@@ -16,6 +18,8 @@ describe('Crypto utilities tests', () => {
 
     const stringifiedData = JSON.stringify(state);
     const key = cryptoUtils.createKey("password", "salt");
+    const testTokenPayload = "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIsImtpZCI6IjFMVE16YWtpaGlSbGFfOHoyQkVKVlhlV01xbyJ9.eyJ2ZXIiOiIyLjAiLCJpc3MiOiJodHRwczovL2xvZ2luLm1pY3Jvc29mdG9ubGluZS5jb20vOTEyMjA0MGQtNmM2Ny00YzViLWIxMTItMzZhMzA0YjY2ZGFkL3YyLjAiLCJzdWIiOiJBQUFBQUFBQUFBQUFBQUFBQUFBQUFJa3pxRlZyU2FTYUZIeTc4MmJidGFRIiwiYXVkIjoiNmNiMDQwMTgtYTNmNS00NmE3LWI5OTUtOTQwYzc4ZjVhZWYzIiwiZXhwIjoxNTM2MzYxNDExLCJpYXQiOjE1MzYyNzQ3MTEsIm5iZiI6MTUzNjI3NDcxMSwibmFtZSI6IkFiZSBMaW5jb2xuIiwicHJlZmVycmVkX3VzZXJuYW1lIjoiQWJlTGlAbWljcm9zb2Z0LmNvbSIsIm9pZCI6IjAwMDAwMDAwLTAwMDAtMDAwMC02NmYzLTMzMzJlY2E3ZWE4MSIsInRpZCI6IjkxMjIwNDBkLTZjNjctNGM1Yi1iMTEyLTM2YTMwNGI2NmRhZCIsIm5vbmNlIjoiMTIzNTIzIiwiYWlvIjoiRGYyVVZYTDFpeCFsTUNXTVNPSkJjRmF0emNHZnZGR2hqS3Y4cTVnMHg3MzJkUjVNQjVCaXN2R1FPN1lXQnlqZDhpUURMcSFlR2JJRGFreXA1bW5PcmNkcUhlWVNubHRlcFFtUnA2QUlaOGpZIn0.1AFWW-Ck5nROwSlltm7GzZvDwUkqvhSQpm55TQsmVo9Y59cLhRXpvB8n-55HCr9Z6G_31_UbeUkoz612I2j_Sm9FFShSDDjoaLQr54CreGIJvjtmS3EkK9a7SJBbcpL1MpUtlfygow39tFjY7EVNW9plWUvRrTgVk7lYLprvfzw-CIqw3gHC-T7IK_m_xkr08INERBtaecwhTeN4chPC4W3jdmw_lIxzC48YoQ0dB1L9-ImX98Egypfrlbm0IBL5spFzL6JDZIRRJOu8vecJvj1mq-IUhGt0MacxX8jdxYLP-KUu2d9MbNKpCKJuZ7p8gwTL5B7NlUdh_dmSviPWrw";
+
 
     it('should generate a valid instance', () => {
         expect(cryptoUtils).toBeInstanceOf(CryptoUtils);
@@ -33,5 +37,10 @@ describe('Crypto utilities tests', () => {
 
         expect(decryptedData).toBe(stringifiedData);
         expect(JSON.parse(decryptedData)).toMatchObject(state);
+    });
+
+    it('should decode a token', () => {
+        const decodedToken = cryptoUtils.decodeAuthToken(testTokenPayload);
+        expect(decodedToken.payload).toMatchObject(TEST_CONSTANTS.ID_TOKEN_CLAIMS);
     });
 });
